feat: reorganize preferences components and styles
