+++ app/assets/javascripts/components/preferences/mock-state.ts
