+++ app/assets/javascripts/components/preferences/view.tsx
@@ -10,11 +10,11 @@ interface PreferencesViewProps {
 export const PreferencesView: FunctionComponent<PreferencesViewProps> = ({
   close,
 }) => (
-  <div className="preferences-view">
-    <TitleBar className="title-bar">
+  <div className="sn-full-screen flex flex-col bg-contrast z-index-preferences">
+    <TitleBar className="items-center justify-between">
       {/* div is added so flex justify-between can center the title */}
       <div className="h-8 w-8" />
-      <Title className="title">Your preferences for Standard Notes</Title>
+      <Title className="text-lg">Your preferences for Standard Notes</Title>
       <IconButton
         onClick={() => {
           close();
