+++ app/assets/stylesheets/_main.scss
@@ -240,3 +240,7 @@ $footer-height: 32px;
 .icon {
   margin-right: 4px;
 }
+
+.z-index-preferences {
+  z-index: $z-index-preferences;
+}
